requirements.txt
autopep8==1.5.7
azure-core==1.27.1
azure-storage-blob==12.16.0
bleach==4.1.0
boto3==1.18.41
botocore==1.21.41
cachetools==5.3.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.5
click==8.0.1
colorama==0.4.4
cryptography==3.4.8
distlib==0.3.6
docutils==0.17.1
fernet==1.0.1
filelock==3.12.2
google-api-core==2.11.1
google-api-python-client==2.91.0
google-auth==2.21.0
google-auth-httplib2==0.1.0
google-cloud==0.34.0
google-cloud-core==2.3.2
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
httplib2==0.22.0
hurry==1.1
hurry.filesize==0.9
idna==3.2
importlib-metadata==4.8.1
install==1.3.5
isodate==0.6.1
jeepney==0.7.1
jmespath==0.10.0
keyring==23.2.1
markdown-it-py==3.0.0
mdurl==0.1.2
mkdir==2020.12.3
orderdict==2020.12.3
packaging==21.0
pkginfo==1.7.1
platformdirs==3.5.3
protobuf==4.23.3
pyaes==1.6.1
pyaml==21.8.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.7.0
pycparser==2.20
pyfiglet==0.8.post1
Pygments==2.10.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==5.4.1
readme-renderer==29.0
requests==2.26.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
rich==13.4.2
rsa==4.9
s3transfer==0.5.0
SecretStorage==3.3.1
setupcfg==2020.12.3
setuppy-generator==2020.12.2
six==1.16.0
toml==0.10.2
tqdm==4.62.2
twine==3.4.2
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.6
values==2020.12.3
virtualenv==20.23.1
webencodings==0.5.1
write==2020.12.3
zipp==3.5.0
